Carved example out of main file at Emil's request
